cdct_framework/requirements.txt
anthropic==0.74.0
anyio==4.2.0
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
cachetools==6.2.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.3.3
cryptography==41.0.7
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
docstring_parser==0.17.0
fonttools==4.60.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gyp==0.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.4
httpx==0.26.0
idna==3.6
Jinja2==3.1.2
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==1.4.9
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
netaddr==0.8.0
numpy==2.3.4
oauthlib==3.2.2
olefile==0.46
openai==2.6.1
packaging==25.0
pandas==2.3.3
pexpect==4.9.0
pillow==10.2.0
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.25.1
pycups==2.0.1
pydantic==1.10.14
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.2.1
pytz==2024.1
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==68.1.2
six==1.16.0
sniffio==1.3.0
systemd-python==235
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.0.7
wadllib==1.3.6
wheel==0.42.0
xdg==5
xkit==0.0.0
